+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handler/HandlerKey.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.handler;
 
 import com.interface21.web.bind.annotation.RequestMethod;
